+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # MiniReelinator
 
+* *[v1.31.0.rc1]*
+  * [FEATURE]: Add support for Kaltura videos
+
 ## v1.30.1 (December 18, 2015)
 * *[v1.30.1.rc3]*
   * [DESIGN]: Moved flight dates underneath the budget section
@@ -21,7 +24,7 @@
 * *[/v1.30.0.rc2]*
 
 * *[v1.30.0.rc1]*
-  * [FEATURE] Add support for Brightcove videos
+  * [FEATURE]: Add support for Brightcove videos
   * Prepare Campaign Manager and Selfie for the removal of Customer entity
 * *[/v1.30.0.rc1]*